Adjusted backend app and added test projects

+++ Backend/CalculationService/Model/ElecricityTariff.cs
@@ -6,7 +6,7 @@ namespace CalculationService.Models
     {
         public string Provider { get; set; }
         public string Name { get; set; }
-        public byte Type { get; set; }
+        public int Type { get; set; }
         public double BaseCost { get; set; }
         public double AdditionalKwhCost { get; set; }
         public int IncludedKwh { get; set; }
