+++ Backend/ElectricityServiceApi/Models/ElectricityTariff.cs
@@ -7,7 +7,7 @@ namespace ElectricityServiceApi.Models
         // new { Provider = providerName, Name = "Product A", Type = 1, BaseCost = 5, AdditionalKwhCost = 22 },
         public string Provider { get; set; }
         public string Name { get; set; }
-        public byte Type { get; set; }
+        public int Type { get; set; }
         public double BaseCost { get; set; }
         public double AdditionalKwhCost { get; set; }
         public int IncludedKwh { get; set; }
@@ -17,7 +17,7 @@ namespace ElectricityServiceApi.Models
             
         }
 
-        public ElectricityTariff(string provider, string name, byte type, double baseCost, double additionalKwhCost, int includedKwh = 0)
+        public ElectricityTariff(string provider, string name, int type, double baseCost, double additionalKwhCost, int includedKwh = 0)
         {
             Provider = provider;
             Name = name;
